Working on op codes
Big update.
Op codes dealing with loading from memory and loading constants are now
working. Breaking also works. Had to switch up some of my program
counter logic as I didn't realize how important counting is because of
the possibility of assuming a '00' for memory reference can be
interpreted as a break statement.