GladOS updated
This is my final commit on time for project 2. Sadly I could not get the BNE operation to work right. I'm going to come to you next week for help in what might be going wrong. I'll continue to work on it regardless, and if I get it working tomorrow or so I will push it to project 2 (with a commit acknowledging its late).